requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.3.0
ccxt==4.4.72
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dateparser==1.2.1
dotenv==0.9.9
fonttools==4.57.0
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.3.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.22.0
pyparsing==3.2.3
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
ta==0.11.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
websockets==15.0.1
yarl==1.18.3
